accent-coach/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
faster-whisper>=1.0.0
pydub==0.25.1
numpy>=1.26.0
TTS>=0.21.3,<0.23
aiofiles==23.2.1
whisperx==3.1.1
torch>=2.2.0
torchaudio>=2.2.0
python-pptx==0.6.23
pypdf==4.2.0
mediapipe==0.10.14
opencv-python==4.10.0.84
librosa==0.10.2.post1
scipy==1.12.0
soundfile==0.12.1
pronouncing==0.2.0
allosaurus==1.0.2
